Cambios uwu

+++ Api/src/modulos/categorias/controlador.js
@@ -1,4 +1,5 @@
 const Tabla = 'tb_categorias';
+const idTabla = 'id_categoria'; 
 const fs = require('fs');
 
 module.exports = function (dbInyectada) {
@@ -15,7 +16,7 @@ module.exports = function (dbInyectada) {
     }
 
     async function uno(id) {
-        const categorias = await db.uno(Tabla, id);
+        const categorias = await db.uno(Tabla, id, idTabla);
         if (categorias) {
             return {
                 ...categorias,
@@ -26,11 +27,11 @@ module.exports = function (dbInyectada) {
     }
 
     async function eliminar(id) {
-        const categorias = await db.uno(Tabla, id);
+        const categorias = await db.uno(Tabla, id, idTabla);
         if (categorias && categorias[0].imagen_categoria) {
             fs.unlinkSync(categorias[0].imagen_categoria);
         }
-        return db.eliminar(Tabla, id, 'id_categoria');
+        return db.eliminar(Tabla, id, idTabla);
     }
 
     async function agregar(data, filePath) {
@@ -43,22 +44,23 @@ module.exports = function (dbInyectada) {
 
     async function actualizar(id, data) {
         const { nombre_categoria, imagen_categoria } = data;
-        const categoria = await db.uno(Tabla, id);
-
-        if (categoria && categoria[0].imagen_categoria && imagen_categoria && imagen_categoria !== categoria[0].imagen_categoria) {
+        const categoria = await db.uno(Tabla, id, idTabla);
+    
+        if (categoria && categoria[0] && categoria[0].imagen_categoria && imagen_categoria && imagen_categoria !== categoria[0].imagen_categoria) {
             // Eliminar la foto anterior del servidor
             fs.unlinkSync(categoria[0].imagen_categoria);
         }
     
         // Actualizar la información del cliente en la base de datos
-        const newData = { id_categoria, nombre_categoria };
-        if (ruta_imagen) {
-            newData.ruta_imagen = ruta_imagen;
+        const newData = { nombre_categoria };
+        if (imagen_categoria) {
+            newData.imagen_categoria = imagen_categoria;
         }
-        const resultado = await db.actualizar(Tabla, newData, 'id_categoria');
+        const resultado = await db.actualizar(Tabla, newData, idTabla, id);
     
         return resultado;
     }
+
